Lightmapping: No lights fix
Currently we quit before baking if no lights are present in the scene.
This is not ideal, as it is fine to bake with only emissive meshes as
well. This patch makes it so that we early fail baking only if there are
no lights nor any emissive meshes.

Change-Id: I74da12da11a8c5cc3e494a2675a341387563576c
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>
(cherry picked from commit 1aa5b242ba44c9d95a4649ddae264a012d1bc816)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>