DefaultMaterial Shader Generator cleanup
This change is a cleanup of how shaders are generated for Materials.

Most of the changes relate to adding more structure and organization
to how the shader code is generated to standardize what is done for
each property and pass. The goal was to make this code more managable
and enable us to more easily add user defined passes.

We dynamically generate some names for image uniforms and coordinates
depending on what properties are set on the material and the referenced
textures, but this code was a bit too ad-hoc so now there is a central
helper to keep track of the image/sampler state and assist with
generation of these dynamic names in a single place.

We also now no longer use the list of images in shader generation.
It should be enough to use the shader kep properties, since that tracks
all necessary maps anyway.

The "Shader Poker" manual test has been updated to include the current
possible keys/features and should now work well with setting image/map
properties.

Change-Id: I65f6eeb872138cae8ac475cf7e761255bcbbde05
Reviewed-by: Christian Strømme <christian.stromme@qt.io>