Fix weird things happening due to data race in effects
Data structures on the render thread cannot reference data on the gui
thread, like they do with the effect commands. Changing a property,
such as the size multiplier in a Buffer then writes the size multiplier
in the command, while the render thread, that works in parallel at that
point may or may not see the changed value depending on the timing.

The expected behavior is that it should not see the new value at all
until the next frame.

Instead, now it can well happen that the render thread sees different
values while preparing an effect for the current frame, which leads to
generating erroneous QRhi command sequences. (such as, recreating a
QRhiTexture that's been used in some already recorded commands in
the current frame)

Avoid this by always making owned copies of commands in
QSSGRenderEffect. In fact this is already the case for all but three
types of commands. Make it happen for those three too. This follows
the general Qt Quick pattern: data must be copied in the synchronize
phase (updatePaintNode/updateSpatialNode), when the gui thread is
blocked, and no gui thread data must be referenced afterwards,
during rendering, since the gui thread moves on at that stage
in parallel with the render thread.

Then, this exposes a second problem: changes to command data
(due to changing Buffer properties for example) do not lead to
rebuilding the effect (in QQuick3DEffect's updateSpatialNode),
i.e. no new copies of the potentially changed data area made.

Previously this worked by accident basically, because e.g. a changed
size multiplier got picked up by the render thread eventually
after the gui thread wrote to the variable. With proper copies of
the data, that won't work.

Rather, Buffer must have a changed signal that leads to rebuilding
the effect in sync, similarly to what happens what Pass emits
a changed signal when its Shaders change. Due to the lose
relationship between Buffers and Effect this is somewhat
challenging, and is for now done by connecting with
UniqueConnection when encountering a Buffer used in
relation with the current effect within the command list.

A few other objects still won't have proper support for this.
There is nothing saying one cannot dynamically change the value
of a BufferInput.sampler property for example. It is unlikely
in practice. These will continue to not notify, and such
dynamic changes will not be correctly picked up by the Effect,
but that is nothing new.

Fixes: QTBUG-140107
Change-Id: If9df5f9fee5a1cb10c087242781b971096560b2c
Reviewed-by: Andy Nichols <andy.nichols@qt.io>