Fix how the Texture Provider Extensions are collected
Previously the texture providers where collected the same way as we do
for resource loaders, that's not necessarily what we want to do as a
Texture provider should be processes for each view unlike the resource
loader which processes it's resource per window. This gives us a bit
more flexibility down the road if we want to enable getting results from
a different view for example.
This change also fixes an issue that would cause the whole extension
list to be re-evaluated each frame if there was a texture provider
extension in the scene.

This change amends 0cb1e39dcd4e687335d69ff209a215923c9270f6

Change-Id: Ic3c9c5a10aed409738a0ae47e44b2fea68f3fd0d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>