QSSGBufferManager::loadMeshBVH: index format fix
According to the documentation and the code, only u16 and u32 are
supported for indexBufferFormat (i16 and i32 are supported by simply
treating them as unsigned). This patch fixes the support for u16 and
returns nullptr if the index is not a 16-bit or 32-bit integer.

Fixes: QTBUG-140780
Pick-to: 6.8 6.5
Change-Id: I47f8a294074acfb6e01f3c28d35d8c8a57abbbad
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 586d1fcd4b0bd6cc4ef99eb9ff29d8788eab7560)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>