Handle Qt Quick Content with pre-multiplied alpha
When a Texture uses sourceItem, if that content contains alpha then
it is safe to assume that the colors will have pre-multipled alpha,
which we need to account for in fragment operations given the used
blend mode. This is done for baseColor/albedo/diffuse textures as
this is the most common place these 4 component textures are relevant.

Task-number: QTBUG-138065
Change-Id: I904bc40d94ac1bc6c9e26cf0ab30309397bd00e3
Reviewed-by: Christian Strømme <christian.stromme@qt.io>