Make sure we track scene changes for source items
In case of a scene change event we need to assume that any layer
(or resource) we acquire will invalidated. With this change we now
release the acquired layer for source items when the window changes,
e.g., when an Item becomes parentless.

Pick-to: 6.8
Fixes: QTBUG-140693
Change-Id: Ib4af5d1a012e8ff352b135799a2e393fbed11f3d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 35583ef9832afcbb3d97872eb5101f1f54025b3a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>