Only do specular on directly lit surfaces
Rewrites the shadow mapping shader a bit to return a ShadowResult that
contains two variables: illumination and isLit. 'isLit' is then used to
do specular only when the surface is directly lit by the current light.
'illumination' is simply how illuminated the surface is while taking the
shadow factor into account.

Change-Id: I38750600f726639ddf3765c4ccbc7650b491543d
Reviewed-by: Kristoffer Skau <kristoffer.skau@qt.io>