Shadows: Try to reduce branching as much as possible
The excessive branching hurt performance, and made the compile times of
D3D11 shaders unacceptably long.

It is now only possible to have 1 shadow softness value for any given
material. This is normally defined by each light, but now only the max
quality for each light affecting a material will be used. This means
that we can hardcode for that material the method of shading shadows
without having to branch or loop based on a uniform value.

Change-Id: Ia574889a0fc5e64ac535a3d52e609ac27fab3ab1
Reviewed-by: Christian Strømme <christian.stromme@qt.io>