Some hardening fixes for the import scene
There's some cases that used to work with import scene that stopped
working after internal rework in 6.10, as we discarded some of those
cases as unsupported. With this change the old behavior should be more
or less restored. Extra comments in the code and printed warnings are
added to let the user know when unsupported usage, like cross window
import scenes, are used.

Pick-to: 6.10
Change-Id: If7daeae1fb233dd4065414210f7137a70d0abe45
Reviewed-by: Andy Nichols <andy.nichols@qt.io>