Doc: Add some notes about limitations of import scenes
As with QtQuick, QtQuick3D does not support sharing items across windows.
While there are cases where this does work today, that should be
considered as accidental. The same goes for subtrees of scenes used in a
different view, so a note is added for that too.

Pick-to: 6.8
Change-Id: I7c03cb17f6ff075479fbb29df3d8a9b6828c8d44
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>
Reviewed-by: Inho Lee <inho.lee@qt.io>
(cherry picked from commit f2bc23742717c002c8599ec6f13694298635d6ac)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>