Fix tonemapping for SkyBoxCubeMap
The cubemap shader would just copy the pixel values directly, which
would give wrong results for sRGB cube maps in linear color space,
as well as for linear cube maps in sRGB color space. This patch adds
a new version of the shader that covers all the different combinations
of incoming pixel formats and tonemap modes. The original shader,
without conditionals, is still used for the case where the color
doesn't need to be modified.

Fixes: QTBUG-141617
Pick-to: 6.10
Change-Id: Idcccfe693dd0e98f644179b4cec07d5bf67aa368
Reviewed-by: Andy Nichols <andy.nichols@qt.io>