Add blue noise randomness to PCF soft shadows
The shader is mostly the same except we sample noise from a 64x64
texture based on the screen position of the fragment. This noise is then
used to rotate the poisson disk/sphere to get more randomness in the
sampling and by extension smoother soft shadows. Using blue noise makes
the noise a bit more uniform and less noticable even though it is still
there even with 16 PCF samples.

[ChangeLog][Third-Party Code] Add CC0-licensed blue noise texture

Change-Id: I5b2a1f7a8cab2189e1ffa5a2fd6d72b25e5cf6b7
Reviewed-by: Andy Nichols <andy.nichols@qt.io>