Add a dedicated type id for the scene root node
Each view has its own scene root, which might be nodes in a larger scene
graph. There are cases where we need to be able to identify those nodes,
so they should have their own type, instead of just being a basic
'node'. For example the recent warning about importing sub parts of a
scene not being supported, we can make it less strict by checking if the
parent node is a scene root, in which case we can avoid printing a
warning, since there's not going to be competing global state.

Change-Id: I50a051a4dbcf4408fe7f6f6aba136b80c6b91cb4
Reviewed-by: Andy Nichols <andy.nichols@qt.io>