Fix crash when rendering using direct mode and oit
The renderer reads back node count from the oit counter image in
OITRenderPass implementation and adds it into resource update.
This is then supposed to be added to the end of render pass
resource updates outside the OITRenderPass. This doesn't happen
when using direct mode so the resource updates keep accumulating.

This removes the need to handle the resource update outside the
OITRenderPass and simply adds the resource update in next prepare
phase.

Pick-to: 6.11
Change-Id: I43336471b54fb34dad1b461f77902e274117fcd1
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>