Remove warning about missing texture in the RenderOutputProvider
It's very possible that the requested render result hasn't been provided
yet, so printing a warning for the first frame doesn't make a whole lot
of sense. An alternative fix would be to change the stage for the
texture provider, but in practice there's no big benefit of managing
extra state for that.
With this change we also reset the dirty state if the source type is set
to 'none'.

Pick-to: 6.11
Change-Id: If467fd2b822e81ec107195178f3e7ca8b6f8c7e6
Reviewed-by: Andy Nichols <andy.nichols@qt.io>