Further refinement of the auto registering of extensions
Remove duplicated code used for the texture provider extension, as that
code can just use the generalized path as well. With this change we also
now put all 'standalone' extensions into the texture provider pass
stage, as that's in essence what 'standalone' means. If pre/post color
is set is therefor only taken into account if the mode is 'Main'.

Pick-to: 6.11
Change-Id: I59c148a7d62ec193ce2a2a10bc4353c4670bbf02
Reviewed-by: Andy Nichols <andy.nichols@qt.io>